# Offload Bitnet token embeddings to the GPU